refactor(auth): rename support package to constant
